first assignment
I would say the key area for me would be the use of the paragraph tag. Adding a paragraph around the primary text of the quote is appropriate as a block of content. Putting that paragraph instead in the citation isn't really necessary however since it's already defined as a citation. What I do see as optional however would be the footer definition done in the video for the citation. I think defining that area as a footer could be argued. So combining both ideas you could probably settle on something below and be okay:

<blockquote>
  <p>"I've attended Front End DevConf 5 times. I look forward to another intensive day of learning and sharing!"</p>
  <cite>- Cody McCode</cite>
</blockquote>